Decouples routing logic and uses express router
