+++ server/config/dev.toml
@@ -1,5 +1,5 @@
 [database]
-database = "todo"
+database = "toodledoodledoo"
 user = "postgres"
 host = "localhost"
 port = 5432
