+++ server/routes/index.js
@@ -1,0 +1,9 @@
+import express from "express";
+import setupTodoRoutes from "./todo.js"
+
+const router = express.Router();
+
+// Initialize /todo routes
+setupTodoRoutes(router);
+
+export default router;
